Remove openbabel dependency in qchem (#2415)
