+++ .github/workflows/tests.yml
@@ -173,7 +173,6 @@ jobs:
 
       - name: Install dependencies
         run: |
-          sudo apt-get install -y openbabel
           conda install psi4 psi4-rt python=3.7 -c psi4
           pip uninstall h5py -y
           pip install pytest pytest-cov pytest-mock flaky h5py==3.2.1
