+++ doc/releases/changelog-dev.md
@@ -101,6 +101,8 @@
     [(#2385)](https://github.com/PennyLaneAI/pennylane/pull/2385)
   - The 6-31G basis set is added to the qchem basis set repo
     [(#2372)](https://github.com/PennyLaneAI/pennylane/pull/2372)
+  - The dependency on openbabel is removed
+    [(#2415)](https://github.com/PennyLaneAI/pennylane/pull/2415)
 
 * Development of a circuit-cutting compiler extension to circuits with sampling
   measurements has begun:
