+++ docker/qchem.dockerfile
@@ -17,7 +17,6 @@ FROM pennylane/base:latest
 # Update and install Qchem
 RUN DEBIAN_FRONTEND="noninteractive" apt-get install tzdata # need to perform this again
 RUN apt-get update \
-    && apt-get -y install --no-install-recommends make git openbabel \
     && rm -rf /var/lib/apt/lists/*
 
 # Setup and Build Qchem
