+++ qchem/pennylane_qchem/qchem/__init__.py
@@ -18,14 +18,9 @@ from .openfermion_obs import (
     one_particle,
     two_particle,
     dipole_of,
-    read_structure,
     meanfield,
-    active_space,
     decompose,
     molecular_hamiltonian,
-    excitations,
-    hf_state,
-    excitations_to_wires,
 )
 from .basis_set import BasisFunction, atom_basis_data, mol_basis_data
 from .convert import import_operator
@@ -61,3 +56,4 @@ from .molecule import Molecule
 from .observable_hf import fermionic_observable, qubit_observable, jordan_wigner, simplify
 from .particle_number import particle_number
 from .spin import spin2, spinz
+from .structure import read_structure, active_space, excitations, hf_state, excitations_to_wires
