--- qchem/tests/qchem_tests/test_ref_files/gdb3.mol5.PDB
@@ -1,22 +1,0 @@
-COMPND    gdb3.mol5.xyz 
-AUTHOR    GENERATED BY OPEN BABEL 2.3.2
-HETATM    1  C   LIG     1       0.361  -0.452  -0.551  1.00  0.00           C  
-HETATM    2  C   LIG     1      -0.714   0.125   0.327  1.00  0.00           C  
-HETATM    3  N   LIG     1       0.683   0.133   0.745  1.00  0.00           N  
-HETATM    4  H   LIG     1       0.442  -1.529  -0.619  1.00  0.00           H  
-HETATM    5  H   LIG     1       0.672   0.102  -1.428  1.00  0.00           H  
-HETATM    6  H   LIG     1      -1.364  -0.560   0.857  1.00  0.00           H  
-HETATM    7  H   LIG     1      -1.149   1.080   0.060  1.00  0.00           H  
-HETATM    8  H   LIG     1       1.093   1.063   0.636  1.00  0.00           H  
-CONECT    1    5    4    2    3                                       
-CONECT    1                                                           
-CONECT    2    1    7    3    6                                       
-CONECT    2                                                           
-CONECT    3    1    2    8                                            
-CONECT    4    1                                                      
-CONECT    5    1                                                      
-CONECT    6    2                                                      
-CONECT    7    2                                                      
-CONECT    8    3                                                      
-MASTER        0    0    0    0    0    0    0    0    8    0    8    0
-END
