+++ qchem/tests/qchem_tests/test_ref_files/gdb3.mol5.XYZ
@@ -1,5 +1,5 @@
 8
-gdb3.mol5.PDB
+gdb3.mol5.XYZ
 C          0.36100       -0.45200       -0.55100
 C         -0.71400        0.12500        0.32700
 N          0.68300        0.13300        0.74500
