--- qchem/tests/qchem_tests/test_ref_files/h2.SDF
@@ -1,8 +1,0 @@
-in Angstrom
- OpenBabel10081915113D
-
-  2  0  0  0  0  0  0  0  0  0999 V2000
-    0.0000    0.0000   -0.3500 H   0  0  0  0  0  0  0  0  0  0  0  0
-    0.0000    0.0000    0.3500 H   0  0  0  0  0  0  0  0  0  0  0  0
-M  END
-$$$$
